+++ tests/test_clean.ts
@@ -45,22 +45,25 @@ describe('Cleaner', () => {
   });
 
   it('should clean two jobs from the queue', async () => {
-    await queue.add('test', { some: 'data' });
-    await queue.add('test', { some: 'data' });
-
     const worker = new Worker(queueName, async job => {}, { connection });
     await worker.waitUntilReady();
 
     worker.run();
+    const completing = new Promise<void>(resolve => {
+      worker.on(
+        'completed',
+        after(2, async () => {
+          const jobs = await queue.clean(0, 0);
+          expect(jobs.length).to.be.eql(2);
+          resolve();
+        }),
+      );
+    });
 
-    queue.on(
-      'completed',
-      after(2, async () => {
-        const jobs = await queue.clean(0, 0);
-        expect(jobs.length).to.be.eql(2);
-      }),
-    );
+    await queue.add('test', { some: 'data' });
+    await queue.add('test', { some: 'data' });
 
+    await completing;
     await worker.close();
   });